TP2/requirements.txt
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
Pillow==10.1.0
selenium==4.15.2
requests==2.31.0